Split up io.cpp
