+++ src/game/editor/editor.cpp
@@ -7824,4 +7824,97 @@ void CEditor::LoadCurrentMap()
 	MapView()->SetWorldOffset(Center);
 }
 
+bool CEditor::Save(const char *pFilename)
+{
+	// Check if file with this name is already being saved at the moment
+	if(std::any_of(std::begin(m_WriterFinishJobs), std::end(m_WriterFinishJobs), [pFilename](const std::shared_ptr<CDataFileWriterFinishJob> &Job) { return str_comp(pFilename, Job->GetRealFileName()) == 0; }))
+		return false;
+
+	return m_Map.Save(pFilename);
+}
+
+bool CEditor::HandleMapDrop(const char *pFileName, int StorageType)
+{
+	return m_Map.HandleMapDrop(pFileName, IStorage::TYPE_ALL_OR_ABSOLUTE);
+}
+
+bool CEditor::Load(const char *pFileName, int StorageType)
+{
+	const auto &&ErrorHandler = [this](const char *pErrorMessage) {
+		ShowFileDialogError("%s", pErrorMessage);
+		Console()->Print(IConsole::OUTPUT_LEVEL_STANDARD, "editor/load", pErrorMessage);
+	};
+
+	Reset();
+	bool Result = m_Map.Load(pFileName, StorageType, std::move(ErrorHandler));
+	if(Result)
+	{
+		str_copy(m_aFileName, pFileName);
+		SortImages();
+		SelectGameLayer();
+		MapView()->OnMapLoad();
+	}
+	else
+	{
+		m_aFileName[0] = 0;
+		Reset();
+	}
+	return Result;
+}
+
+bool CEditor::Append(const char *pFileName, int StorageType)
+{
+	CEditorMap NewMap;
+	NewMap.m_pEditor = this;
+
+	const auto &&ErrorHandler = [this](const char *pErrorMessage) {
+		ShowFileDialogError("%s", pErrorMessage);
+		Console()->Print(IConsole::OUTPUT_LEVEL_STANDARD, "editor/append", pErrorMessage);
+	};
+	if(!NewMap.Load(pFileName, StorageType, std::move(ErrorHandler)))
+		return false;
+
+	// modify indices
+	static const auto &&s_ModifyAddIndex = [](int AddAmount) {
+		return [AddAmount](int *pIndex) {
+			if(*pIndex >= 0)
+				*pIndex += AddAmount;
+		};
+	};
+	NewMap.ModifyImageIndex(s_ModifyAddIndex(m_Map.m_vpImages.size()));
+	NewMap.ModifySoundIndex(s_ModifyAddIndex(m_Map.m_vpSounds.size()));
+	NewMap.ModifyEnvelopeIndex(s_ModifyAddIndex(m_Map.m_vpEnvelopes.size()));
+
+	// transfer images
+	for(const auto &pImage : NewMap.m_vpImages)
+		m_Map.m_vpImages.push_back(pImage);
+	NewMap.m_vpImages.clear();
+
+	// transfer sounds
+	for(const auto &pSound : NewMap.m_vpSounds)
+		m_Map.m_vpSounds.push_back(pSound);
+	NewMap.m_vpSounds.clear();
+
+	// transfer envelopes
+	for(const auto &pEnvelope : NewMap.m_vpEnvelopes)
+		m_Map.m_vpEnvelopes.push_back(pEnvelope);
+	NewMap.m_vpEnvelopes.clear();
+
+	// transfer groups
+	for(const auto &pGroup : NewMap.m_vpGroups)
+	{
+		if(pGroup != NewMap.m_pGameGroup)
+		{
+			pGroup->m_pMap = &m_Map;
+			m_Map.m_vpGroups.push_back(pGroup);
+		}
+	}
+	NewMap.m_vpGroups.clear();
+
+	SortImages();
+
+	// all done \o/
+	return true;
+}
+
 IEditor *CreateEditor() { return new CEditor; }
